content/base/test/test_bug320799.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=320799
-->
<head>
  <title>Test for Bug 320799</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=320799">Mozilla Bug 320799</a>
<p id="display">
  <select id="s" style="width: 100px; -moz-box-sizing: padding-box">
    <option>This is a test, it really is a test I tell you</option>
  </select>
  <select id="s2">
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
    <option>x</option>
  </select>
  <select id="s3">
    <option>x</option>
  </select>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 320799 **/
is($("s").scrollWidth, 100, "Scroll width should not include dropdown contents");
is($("s2").clientWidth, $("s3").clientWidth,
   "Client width should not depend on the dropdown's vertical scrollbar");
</script>
</pre>
</body>
</html>
